third_party/blink/web_tests/media/media-fragments/TC0024-expected.txt


Title: TC0024
Fragment: 't=npt:3,7'
Comment: equivalent to #t=3,7
EVENT(canplaythrough)
EXPECTED (video.currentTime == '3') OK
RUN(video.play())
EVENT(pause)
EXPECTED (video.currentTime - fragmentEndTime <= '0.5') OK
END OF TEST
